+++ client/src/components/shared/Navbar.tsx
